CategoryPlot
Support for a default colors
